+++ webui/src/locales/cs.js
@@ -15,6 +15,7 @@ export default {
     ok: 'OK',
     unknownError: 'Neznámá chyba',
     copyFailed: 'Kopírování selhalo',
+    clear: 'Vymazat',
     network: {
       serverErrorTitle: 'Chyba serveru',
       serverError: 'Chyba serveru: {status}',
@@ -295,7 +296,7 @@ export default {
     otaSuccess: 'Stažení úspěšné. Instalace...',
     currentVersion: 'Současná verze',
     installedVersion: 'Nainstalovaná verze',
-    versionInfo: 'Modernizovaný fork v2.2.3-Beta.31 od Xerolux (2025) - Založeno na původní práci Alexandra Reinerta.',
+    versionInfo: 'Modernizovaný fork v2.2.3-Beta.32 od Xerolux (2025) - Založeno na původní práci Alexandra Reinerta.',
     updateAvailable: 'Aktualizace k dispozici!',
     updateAvailableMessage: 'Je k dispozici aktualizace na verzi {latestVersion}.',
     upToDate: 'Firmware je aktuální',