+++ webui/src/locales/nl.js
@@ -15,6 +15,7 @@ export default {
     ok: 'OK',
     unknownError: 'Onbekende fout',
     copyFailed: 'Kopiëren mislukt',
+    clear: 'Wissen',
     network: {
       serverErrorTitle: 'Serverfout',
       serverError: 'Serverfout: {status}',
@@ -295,7 +296,7 @@ export default {
     otaSuccess: 'Download succesvol. Installeren...',
     currentVersion: 'Huidige Versie',
     installedVersion: 'Geïnstalleerde versie',
-    versionInfo: 'Gemoderniseerde fork v2.2.3-Beta.31 door Xerolux (2025) - Gebaseerd op het originele werk van Alexander Reinert.',
+    versionInfo: 'Gemoderniseerde fork v2.2.3-Beta.32 door Xerolux (2025) - Gebaseerd op het originele werk van Alexander Reinert.',
     updateAvailable: 'Update beschikbaar!',
     updateAvailableMessage: 'Een update naar versie {latestVersion} is beschikbaar.',
     upToDate: 'Firmware is up-to-date',