+++ webui/src/locales/no.js
@@ -15,6 +15,7 @@ export default {
     ok: 'OK',
     unknownError: 'Ukjent feil',
     copyFailed: 'Kopiering mislyktes',
+    clear: 'Tøm',
     network: {
       serverErrorTitle: 'Serverfeil',
       serverError: 'Serverfeil: {status}',
@@ -295,7 +296,7 @@ export default {
     otaSuccess: 'Nedlasting vellykket. Installerer...',
     currentVersion: 'Nåværende Versjon',
     installedVersion: 'Installert versjon',
-    versionInfo: 'Modernisert fork v2.2.3-Beta.31 av Xerolux (2025) - Basert på originalarbeidet til Alexander Reinert.',
+    versionInfo: 'Modernisert fork v2.2.3-Beta.32 av Xerolux (2025) - Basert på originalarbeidet til Alexander Reinert.',
     updateAvailable: 'Oppdatering tilgjengelig!',
     updateAvailableMessage: 'En oppdatering til versjon {latestVersion} er tilgjengelig.',
     upToDate: 'Fastvaren er oppdatert',